beats_store/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.10.1
django-jazzmin==2.6.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
stripe==7.8.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
